Add bls library
